Filter by archived status according to profile
